\name{sybilcycleFreeFlux-package}
\alias{sybilcycleFreeFlux-package}
\alias{sybilcycleFreeFlux}
\docType{package}
\title{
  Find flux distribution free of cycles
}
\description{
The package \code{sybilcycleFreeFlux} implements some ideas to get rid of cycles.
 Also implement sampling and loopless sampling.
}
\details{
\tabular{ll}{
Package: \tab sybilcycleFreeFlux\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-06-03\cr
License: \tab GPL Version 3\cr
LazyLoad: \tab yes\cr
Depends: \tab \code{\link{sybil}}, methods\cr
}
}
\author{Abdelmoneim Amer Desouki}
\keyword{ package }
\seealso{
	\code{\link{sybil}} 
	\code{\link{lrFBA}} 
}